csl-doc-tracker/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
